This commit adds the objective-c sample app.
The objective-c sample app has been built as a rough
approximation of our swift sample app. Where possible,
the same approach has been taken for creating models,
controllers and utility classes. The storyboard should be
almost identical, with some differences found in layout
(and constraints used, most likely). The storyboard was
removed from localization to eliminate some warnings
related to texts in UI elements - this can be reintroduced
in the future if we decide to localize these applications.

The podfile has been added and should mirror the podfile
used in the Swift sample app. The Pods directory and
xcworkspace are excluded from source control as part
of this project's gitignore file. Adding the SDK to the project
using `pod install` was verified before committing these
changes.