Update Sample app based on feedback.
This updates the callback name for the fail to load Native Ad
delegate callback where it was incorrectly maintained. Also,
buttons have been styled with corner radius settings as discussed.
Also updated positioning of buttons to insure that they were able
to be seen with no visible overlapping on smallest supported iOS
devices.